+++ src/Presentation/SmartStore.Web.Framework/DependencyRegistrar.cs
@@ -21,7 +21,7 @@ using SmartStore.Core.Events;
 using SmartStore.Core.Fakes;
 using SmartStore.Core.Infrastructure;
 using SmartStore.Core.Infrastructure.DependencyManagement;
-using SmartStore.Core.IO.Media;
+using SmartStore.Core.IO;
 using SmartStore.Core.IO.VirtualPath;
 using SmartStore.Core.IO.WebSite;
 using SmartStore.Core.Localization;
@@ -775,7 +775,7 @@ namespace SmartStore.Web.Framework
 	{
 		protected override void Load(ContainerBuilder builder)
 		{
-			builder.RegisterType<FileSystemStorageProvider>().As<IStorageProvider>().InstancePerRequest();
+			builder.RegisterType<LocalFileSystem>().As<IFileSystem>().InstancePerRequest();
 			builder.RegisterType<DefaultVirtualPathProvider>().As<IVirtualPathProvider>().InstancePerRequest();
 			builder.RegisterType<WebSiteFolder>().As<IWebSiteFolder>().InstancePerRequest();
 		}